pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hflav-fair-client"
version = "0.1.0"
description = "HFLAV FAIR client - Helper to fetch HFLAV data from Zenodo"
readme = "README.md"
authors = [{ name = "sergio" }]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "requests>=2.28",
  "pydantic>=2.0",
  "genson>=1.3.0",
  "python-dateutil>=2.8",
  "jsonschema>=4.25.1",
  "rich>=14.2.0",
  "dependency-injector>=4.48.2",
  "python-gitlab>=7.0.0",
  "matplotlib>=3.10.7",
  "jsonpath_ng>=1.7.0",
  "requests-cache>=1.2.1",
  "python-dotenv>=1.2.1",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Physics",
]

[project.optional-dependencies]
# Dependencies for testing
test = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "pytest-mock>=3.10",
  "pytest-benchmark>=4.0",
  "memory-profiler>=0.61.0",
]
# Dependencies for development (includes test dependencies)
# Hypothesis is a possibility to generate random data during development and not depend on the published data
dev = [
  "hflav-fair-client[test]",
  "pysonar>=1.3.0.4086",
  "hypothesis-jsonschema>=0.23.1",
  "hypothesis>=6.148.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["hflav_fair_client*"]

[tool.setuptools.package-data]
"hflav_fair_client" = ["py.typed"]

[project.urls]
Homepage = "https://example.org/"
Repository = "https://github.com/yourusername/hflav-fair-client"
Documentation = "https://yourusername.github.io/hflav-fair-client/"
"Bug Tracker" = "https://github.com/yourusername/hflav-fair-client/issues"

# Optional: Configuration for pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--strict-config", "--tb=short"]
markers = [
  "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "performance: marks tests as performance tests (deselect with '-m \"not performance\"')",
]

# Optional: Configuration for coverage
[tool.coverage.run]
source = ["hflav_fair_client"]
omit = [
  "*/tests/*",
  "*/test_*.py",
  "*/__pycache__/*",
  "*/setup.py",
  "*/conftest.py",
  "*/hflav_fair_client/cache.py",
  "*/hflav_fair_client/config.py",
  "*/hflav_fair_client/container.py",
  "*/hflav_fair_client/logger.py",
  "*/hflav_fair_client/main.py",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __",
  "raise NotImplementedError",
  "@property",
  "@abstractmethod",
  "if TYPE_CHECKING:",
  "class .*\\(Protocol\\):",
]
show_missing = true
skip_covered = false
